refactor: callback_handlers
